+++ components/Favorites/FavoriteItemsBar.tsx
@@ -5,42 +5,24 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { Badge, Spinner } from 'react-bootstrap'
 import api from '../../api/ApiHelper'
-import { CUSTOM_EVENTS } from '../../api/ApiTypes.d'
+import { getApiPricesChange } from '../../api/_generated/skyApi'
+import { parsePriceMovement } from '../../utils/Parser/APIResponseParser'
 import { convertTagToName, numberWithThousandsSeparators } from '../../utils/Formatter'
 import { getFavoriteItems } from '../../utils/FavoriteItemUtils'
+import { useFavorites } from './FavoritesContext'
 import FavoriteToggle from './FavoriteToggle'
+import FavoriteItemCard from './FavoriteItemCard'
 import SubscribeButton from '../SubscribeButton/SubscribeButton'
 import styles from './FavoriteItemsBar.module.css'
 
 const REFRESH_INTERVAL = 5 * 60 * 1000
 
 function FavoriteItemsBar() {
-    let [favorites, setFavorites] = useState<FavoriteItemEntry[]>(() => getFavoriteItems())
+    const { favorites } = useFavorites()
     let [priceMovements, setPriceMovements] = useState<Record<string, ItemPriceMovement>>({})
     let [isLoading, setIsLoading] = useState(false)
     let [error, setError] = useState<string | null>(null)
-
-    useEffect(() => {
-        function load() {
-            setFavorites(getFavoriteItems())
-        }
-
-        load()
-
-        function onFavoritesUpdated(event?: Event) {
-            let detailFavorites = (event as CustomEvent)?.detail as FavoriteItemEntry[] | undefined
-            if (Array.isArray(detailFavorites)) {
-                setFavorites(detailFavorites)
-            } else {
-                load()
-            }
-        }
-
-        document.addEventListener(CUSTOM_EVENTS.FAVORITES_UPDATED, onFavoritesUpdated)
-        return () => {
-            document.removeEventListener(CUSTOM_EVENTS.FAVORITES_UPDATED, onFavoritesUpdated)
-        }
-    }, [])
+    // Favorites are provided by FavoritesProvider; no DOM custom events required.
 
     useEffect(() => {
         let cancelled = false
@@ -57,9 +39,14 @@ function FavoriteItemsBar() {
             setIsLoading(true)
             setError(null)
             try {
-                let data = await api.getPriceMovements(favorites.map(f => f.tag))
+                const resp = await getApiPricesChange({ itemTags: favorites.map(f => f.tag) })
+                const raw = resp.data || {}
+                const parsed: Record<string, ItemPriceMovement> = {}
+                Object.keys(raw).forEach(tag => {
+                    parsed[tag] = parsePriceMovement(tag, raw[tag])
+                })
                 if (!cancelled) {
-                    setPriceMovements(data)
+                    setPriceMovements(parsed)
                 }
             } catch (e) {
                 if (!cancelled) {
@@ -117,77 +104,7 @@ function FavoriteItemsBar() {
                 {favorites.map(favorite => {
                     const movement = priceMovements[favorite.tag]
                     const isMovementLoading = isLoading && !movement
-                    const currentPrice = movement?.now
-                    const previousPrice = movement?.recent
-                    const delta: number | null = currentPrice !== undefined && previousPrice !== undefined ? currentPrice - previousPrice : null
-                    let percentChange: number | null = null
-                    if (delta !== null && previousPrice !== undefined && Math.abs(previousPrice) > 1) {
-                        percentChange = (delta / previousPrice) * 100
-                    }
-                    const monthlyPrice = movement?.monthly
-                    const monthlyDelta: number | null = currentPrice !== undefined && monthlyPrice !== undefined ? currentPrice - monthlyPrice : null
-                    const monthlyTooltip =
-                        monthlyDelta !== null
-                            ? `Monthly change: ${monthlyDelta >= 0 ? '+' : ''}${numberWithThousandsSeparators(Math.round(monthlyDelta))} coins`
-                            : isMovementLoading
-                              ? 'Loading monthly comparison…'
-                              : 'Monthly comparison unavailable'
-                    const isPositive = (delta ?? 0) >= 0
-                    const itemName = favorite.name || convertTagToName(favorite.tag)
-                    const item: Item = {
-                        tag: favorite.tag,
-                        name: favorite.name,
-                        iconUrl: favorite.iconUrl,
-                        bazaar: favorite.bazaar
-                    }
-                    const iconSrc = item.iconUrl || api.getItemImageUrl(item)
-                    return (
-                        <article key={favorite.tag} className={styles.card}>
-                            <Link href={`/item/${favorite.tag}`} className={styles.cardLink}>
-                                <div className={styles.iconWrapper}>
-                                    <Image src={iconSrc} alt="" width={40} height={40} />
-                                </div>
-                                <div className={styles.details}>
-                                    <span className={styles.name}>{itemName}</span>
-                                    <div className={styles.stats}>
-                                        <span
-                                            className={`${styles.delta} ${delta !== null ? (isPositive ? styles.positive : styles.negative) : ''} ${
-                                                isMovementLoading ? styles.skeletonInline : ''
-                                            }`}
-                                            title={monthlyTooltip}
-                                        >
-                                            {movement && delta !== null
-                                                ? delta === 0
-                                                    ? 'No change'
-                                                    : `${isPositive ? '+' : ''}${numberWithThousandsSeparators(Math.round(delta))} coins`
-                                                : isMovementLoading
-                                                  ? ''
-                                                  : 'No data'}
-                                            {movement && percentChange !== null && delta !== 0 ? (
-                                                <span className={styles.percent}>
-                                                    {' '}
-                                                    ({`${percentChange >= 0 ? '+' : ''}${percentChange.toFixed(1)}%`})
-                                                </span>
-                                            ) : null}
-                                        </span>
-                                        {movement && currentPrice !== undefined ? (
-                                            <Badge bg="secondary">Now: {numberWithThousandsSeparators(Math.round(currentPrice))}</Badge>
-                                        ) : isMovementLoading ? (
-                                            <span className={`${styles.skeletonInline} ${styles.badgeSkeleton}`} aria-hidden="true"></span>
-                                        ) : null}
-                                    </div>
-                                </div>
-                            </Link>
-                            <div className={styles.actions}>
-                                <SubscribeButton
-                                    topic={favorite.tag}
-                                    type={favorite.bazaar ? 'bazaar' : 'item'}
-                                    buttonContent={<span>Notify</span>}
-                                />
-                                <FavoriteToggle item={item} size="small" />
-                            </div>
-                        </article>
-                    )
+                    return <FavoriteItemCard key={favorite.tag} favorite={favorite} movement={movement} isMovementLoading={isMovementLoading} />
                 })}
             </div>
         </section>
