+++ components/Providers/Providers.tsx
@@ -3,6 +3,7 @@ import { MatomoProvider, createInstance } from '@jonkoops/matomo-tracker-react'
 import { GoogleOAuthProvider } from '@react-oauth/google'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { getQueryClient } from '../../utils/QueryUtils'
+import { FavoritesProvider } from '../Favorites/FavoritesContext'
 
 const matomoTrackingInstance = createInstance({
     urlBase: 'https://track.coflnet.com',
@@ -14,7 +15,9 @@ export function Providers({ children }) {
     return (
         <MatomoProvider value={matomoTrackingInstance}>
             <QueryClientProvider client={queryClient}>
-                <GoogleOAuthProvider clientId="570302890760-nlkgd99b71q4d61am4lpqdhen1penddt.apps.googleusercontent.com">{children}</GoogleOAuthProvider>
+                <GoogleOAuthProvider clientId="570302890760-nlkgd99b71q4d61am4lpqdhen1penddt.apps.googleusercontent.com">
+                    <FavoritesProvider>{children}</FavoritesProvider>
+                </GoogleOAuthProvider>
             </QueryClientProvider>
         </MatomoProvider>
     )
